Issue #175 Idle content-repository pod along with Jenkins
Add content-repository pod to be idled along jenkins.

We didn't do anything complicated here, if we are going to idle jenkins we are
doing content-repository along the way, if user access jenkins.openshift.io and
then get the unidle apicall called we unidle the content-repository too.

We are not handling the scenario when the user is unidling manually, the
content-repository would be left unidled until the next idle round. That's a
bit stretchy scenario for the amount of work.