Fix #241 Rename internal IsIdle method to State that returns a type
OpenshiftClient IsIdle method although sound like it returns a bool
actually returns an int. This method has been renames to return a
type `PodState` which also implements the `stringer` interface so
that the state can be easily (json) marshalled. As a side-effect
of this, the `state` api is simplified a bit.