Fix #233 Send user to idler once per function
This commit makes the HandleBuild() and HandleDeploymentConfig()
functions to send a user to the idler channel exactly once.

Every time the user is sent to the idler channel and the decision
comes out to be idle or unidle, then the respective retry counts
are incremented by 1. However, in the current code the maximum
retry counts allowed in 45 minutes are 3. Sending users to the
idler channel multiple times leads to some requests to fail
because the max retry count is reached.