Issue #163 Could not find condition 'Available'
The error "Error from DC callback: could not find condition 'Available'
" occurs when we don't have any available deployments.If a pod is
coming up the condition would be 'Progressing' and if it fails it would
be 'ReplicaFailure'. To find that there are no available deployments is
an information rather than an error, while watching DeploymentConfigs.
Whenever we see this condition, we just continue the loop and it does
not block us in any way nor does it lead to any erratic behaviour. We
simply wait for the pod to be 'Available'. This error should be logged
as info rather than error