Issue #259 Move to Quay.io
This PR is part of the effort to move to Quay.

This is the list of changes in this PR:

- Pushes to Quay instead of the Devshift registry
- Uses an image hosted in Quay to build the RHEL based container
  image
- Uses the new env-toolkit to load variables from jenkins-env
- Changes the default path of the image to quay (note that this
  should not affect production because it is overridden in the
  saas repo)

A companion PR should have been submitted to the appropriate saas
repo to change the staging url from devshift to quay. The PR to the
saas repo should be merged before this one.
https://github.com/openshiftio/saas-openshiftio/pull/953