Change Idle/UnIdle condition evaluation
This patch changes the Condition.Eval from returning `true|false` to
return on of the 3 states `NoAction, Idle, UnIdle`. This allows certain
conditions to indicate to idler that it does not need to change the
state of Jenkins. E.g NoAction is returned
  - when there is an error evaluating condition
  - When the current build is done but has not gone past the `idleTime`
    based on the last build
  - When dc-condition does not know when the idler was un-idled

The patch also sets `JenkinsLastUpdate` time if idler un-idles Jenkins
and it doesn't know when the last updated time was.

Fixes:
 - https://github.com/openshiftio/openshift.io/issues/4215