Idle users by default when toggle service is unreachable
When connection to feature toggle service is broken, the unleash_toggle
returned that the Idler is disabled for the user. This resulted in idler
not idling certain users. This patch changes the fallback so that by
default all users are idled unless explicitly disabled by feature toggle
service (f8toggles) and the service is reachable.

Fixes: https://github.com/fabric8-services/fabric8-jenkins-idler/issues/333